mpi/init: move MPIR_Init_thread and MPIR_Finalize together
Init and Finalize often need matching code, moving the code into the
same unit makes checking easier. Another motivation is both need manage
init state, init counter and use shared init lock. Having them together
allows use of static variables rather than maintaining these init-only
variables in a global scope.
