+++ src/mpi/init/Makefile.mk
@@ -7,7 +7,6 @@ mpi_core_sources += \
     src/mpi/init/init_impl.c      \
     src/mpi/init/init_util.c      \
     src/mpi/init/mpir_init.c      \
-    src/mpi/init/mpir_finalize.c  \
     src/mpi/init/globals.c        \
     src/mpi/init/initinfo.c       \
     src/mpi/init/local_proc_attrs.c \
